+++ R/remove_MOSAIC_python_env.R
@@ -1,8 +1,8 @@
 #' Remove the MOSAIC Python Environment
 #'
 #' @description
-#' This function checks whether the MOSAIC Python conda environment exists at "~/.MOSAIC_conda_env".
-#' If it exists, it unlinks it from \code{reticulate}, removes the directory, and resets the RETICULATE_PYTHON global variable.
+#' This function checks whether the MOSAIC Python virtual environment exists at "~/.virtualenvs/r-mosaic".
+#' If it exists, it removes the directory and resets the RETICULATE_PYTHON global variable.
 #'
 #' @param force Logical. If TRUE, removes the environment without prompting for confirmation. Default is FALSE.
 #'